Code/

+++ Code/DataStructs/testDatastructs.cpp
@@ -1428,7 +1428,7 @@ TEST_CASE("test14RealVect") {
   }
   RealValueVect vect2(30);
   for (i = 0; i < vect2.getLength(); ++i) {
-    vect2.setVal(i, double(1.0 / (i + 1.0)));
+    vect2.setVal(i, 1.0 / (i + 1.0));
   }
 
   REQUIRE(vect2.getLength() == 30);
